Set up the project and mount point at build time.
I believe it is better to set up the py project and fuse mount point
at build time instead of at runtime. This way we can set it up as root
and chown it to the local user during build time, then we can boot as
the local user and everything should be set up.