Updating the Theia Docker build to match others.
I am updating the Theia Docker build with all the changes done to the
Debian and Alpine Docker builds, with the difference that the Theia
build uses the local theia user. This also sets up sudo correctly in
Alpine. I should probably fix that in the pure Alpine build too.